refactor: 薄いフィルタコンポーネントを整理
- ロジックを持たない DatePresetFilter / ReadStatusFilter は単一箇所でしか
  使われないため filter-controls へインライン化し、コンポーネントを廃止
- MediaFilter は inbox ページでも利用しているため、options 重複を避けて
  コンポーネントのまま維持
- 各所で参照される MediaType / ReadStatusType は use-articles へ移動
